Debugged Program

+++ prob02/main.cpp
@@ -6,7 +6,7 @@ int main()
 {
   const int NUM_INTS = 3; // number of values to be averaged
 
-  float num1, num2, num3;
+  int num1, num2, num3;
   double average; // average of the numbers
 
     // get the numbers
@@ -14,7 +14,7 @@ int main()
     std::cin >> num1 >> num2 >> num3;
 
     //calculate the average of the numbers
-       average = (num1 + num2 + num3) / NUM_INTS;
+       average = double(num1 + num2 + num3) / NUM_INTS;
 
   // display the average, including the decimal portion
   std::cout << "The average of " << num1 << ", " << num2 << ", and " << num3
